flightproject/requirements.txt
aiofiles==0.8.0
aiosqlite==0.17.0
amqp==5.1.1
anyio==3.6.1
async-timeout==4.0.2
atlastk==0.13.2
attrs==22.1.0
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
crochet==2.0.0
cryptography==37.0.4
cssselect==1.1.0
Deprecated==1.2.13
fastapi==0.79.0
filelock==3.7.1
h11==0.13.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iso8601==1.0.2
itemadapter==0.7.0
itemloaders==1.0.4
jmespath==1.0.1
kombu==5.2.4
lxml==4.9.1
packaging==21.3
parsel==1.6.0
prompt-toolkit==3.0.30
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
PyDispatcher==2.0.5
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pypika-tortoise==0.1.6
pytz==2022.1
queuelib==1.6.2
redis==4.3.4
requests==2.28.1
requests-file==1.5.1
Scrapy==2.6.2
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tldextract==3.3.1
tortoise-orm==0.19.2
Twisted==22.4.0
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
wrapt==1.14.1
zope.interface==5.4.0